Add initial setup of router for LodeRunner.UI with NotFoundPage and TestPage
