+++ src/LodeRunner.UI/package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "babel-jest": "^26.6.3",
+    "hookrouter": "^1.2.5",
     "jest": "^26.6.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
