+++ src/LodeRunner.UI/src/components/App/routes.js
@@ -1,0 +1,5 @@
+import TestPage from "../TestPage";
+
+const routes = { "/": () => <TestPage /> };
+
+export default routes;
