+++ src/LodeRunner.UI/src/components/Clients/index.js
@@ -1,13 +1,13 @@
 import { useContext } from "react";
 import PropTypes from "prop-types";
 import CheckMarkIcon from "../CheckMarkIcon";
-import { ClientsContext, DisplayContext } from "../../contexts";
+import { ClientsContext, TestPageContext } from "../../contexts";
 import { CLIENT, CLIENT_STATUSES } from "../../models";
 import { MAIN_CONTENT } from "../../utilities/constants";
 import "./styles.css";
 
 const Clients = ({ setFetchClientsInterval }) => {
-  const { setMainContent } = useContext(DisplayContext);
+  const { setMainContent } = useContext(TestPageContext);
   const {
     clients,
     setSelectedClientIds,
