+++ src/LodeRunner.UI/src/components/NotFoundPage/index.js
@@ -1,0 +1,3 @@
+const NotFoundPage = () => <div>Page Not Found</div>;
+
+export default NotFoundPage;
