+++ src/LodeRunner.UI/src/contexts/app.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const AppContext = createContext();
+
+export default AppContext;
