--- src/LodeRunner.UI/src/contexts/display.js
@@ -1,5 +1,0 @@
-import { createContext } from "react";
-
-const DisplayContext = createContext();
-
-export default DisplayContext;
