+++ src/LodeRunner.UI/src/contexts/testPage.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const TestPageContext = createContext();
+
+export default TestPageContext;
